[Record/CMA-61] 타이머 운동 기록 생성

+++ app/src/main/java/com/climus/climeet/presentation/ui/main/record/timer/data/ClimbingRecordData.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "climbing_record")
 data class ClimbingRecordData(
     @PrimaryKey(autoGenerate = true)
-    val gymId: Int = 0,
+    val gymId: Long = 0,
     val date: String = "",
     val time: String = "",
     val avgDifficulty: Int = 0,
@@ -14,7 +14,7 @@ data class ClimbingRecordData(
 )
 
 data class ClimbingRecord(
-    val routeId: Int = 0,
+    val routeId: Long = 0,
     val attemptCount: Int = 0,
     val isCompleted: Boolean = false
 )
